requirements.txt
python-chess>=1.9.0
torch>=2.5.0
transformers>=4.45.0
scikit-learn>=1.3.2